+++ R/checkFunction.R
@@ -33,7 +33,7 @@ checkFunction = function(x, args = NULL, ordered = FALSE, nargs = NULL, null.ok 
 
   if (!is.null(args)) {
     qassert(args, "S")
-    fargs = names(formals(x)) %??% character(0L)
+    fargs = names(formals(args(x))) %??% character(0L)
 
     if (length(args) == 0L) {
       if (length(fargs) > 0L)
@@ -55,7 +55,7 @@ checkFunction = function(x, args = NULL, ordered = FALSE, nargs = NULL, null.ok 
 
   if (!is.null(nargs)) {
     nargs = asCount(nargs)
-    fnargs = length(setdiff(names(formals(x)) %??% character(0L), "..."))
+    fnargs = length(setdiff(names(formals(args(x))) %??% character(0L), "..."))
     if (nargs != fnargs)
       return(sprintf("Must have exactly %i formal arguments, but has %i", nargs, fnargs))
   }