+++ R/makeAssertion.R
@@ -53,21 +53,22 @@ makeAssertion = function(x, res, var.name, collection) {
 makeAssertionFunction = function(check.fun, c.fun = NULL, include.ns = TRUE, env = parent.frame()) {
   fn.name = if (!is.character(check.fun)) deparse(substitute(check.fun)) else check.fun
   check.fun = match.fun(check.fun)
+  fargs = formals(args(check.fun))
   x = NULL
 
   new.fun = function() TRUE
   if (include.ns) {
-    formals(new.fun) = c(formals(check.fun), alist(.var.name = checkmate::vname(x), add = NULL))
+    formals(new.fun) = c(fargs, alist(.var.name = checkmate::vname(x), add = NULL))
     tmpl = "{ res = %s(%s); checkmate::makeAssertion(x, res, .var.name, add) }"
   } else {
-    formals(new.fun) = c(formals(check.fun), alist(.var.name = vname(x), add = NULL))
+    formals(new.fun) = c(fargs, alist(.var.name = vname(x), add = NULL))
     tmpl = "{ res = %s(%s); makeAssertion(x, res, .var.name, add) }"
   }
 
   if (is.null(c.fun)) {
-    body(new.fun) = parse(text = sprintf(tmpl, fn.name, paste0(names(formals(check.fun)), collapse = ", ")))
+    body(new.fun) = parse(text = sprintf(tmpl, fn.name, paste0(names(fargs), collapse = ", ")))
   } else {
-    body(new.fun) = parse(text = sprintf(tmpl, ".Call", paste0(c(c.fun, names(formals(check.fun))), collapse = ", ")))
+    body(new.fun) = parse(text = sprintf(tmpl, ".Call", paste0(c(c.fun, names(fargs)), collapse = ", ")))
   }
   environment(new.fun) = env
   return(new.fun)