+++ R/makeTest.R
@@ -38,14 +38,15 @@ makeTest = function(res) {
 makeTestFunction = function(check.fun, c.fun = NULL, env = parent.frame()) {
   fn.name = if (!is.character(check.fun)) deparse(substitute(check.fun)) else check.fun
   check.fun = match.fun(check.fun)
+  fargs = formals(args(check.fun))
 
   new.fun = function() TRUE
-  formals(new.fun) = formals(check.fun)
+  formals(new.fun) = fargs
   tmpl = "{ identical(%s(%s), TRUE) }"
   if (is.null(c.fun)) {
-    body(new.fun) = parse(text = sprintf(tmpl, fn.name, paste0(names(formals(check.fun)), collapse = ", ")))
+    body(new.fun) = parse(text = sprintf(tmpl, fn.name, paste0(names(fargs), collapse = ", ")))
   } else {
-    body(new.fun) = parse(text = sprintf(tmpl, ".Call", paste0(c(c.fun, names(formals(check.fun))), collapse = ", ")))
+    body(new.fun) = parse(text = sprintf(tmpl, ".Call", paste0(c(c.fun, names(fargs)), collapse = ", ")))
   }
   environment(new.fun) = env
   return(new.fun)