+++ NEWS.md
@@ -10,6 +10,7 @@
 * New arguments `max.rows` and `max.cols` to check for maximum number
   of rows and columns for matrices, data.frames, tibbles and data.tables.
 * Fixed an error message in `checkChoice`.
+* Fixed `*Function` to work properly with Primitives.
 
 # Version 1.8.5
 * Added `*POSIXct` to check POSIXct data-time objects in POSIXct format.